requirements.txt
tensorflow-gpu==2.11.0
tensorflow-hub==0.12.0
grpcio-tools==1.48.2
fire
pandas
tqdm
black
pytest
pytest-mock
mypy
types-protobuf